+++ nextgisweb/feature_attachment/nodepkg/export-form/index.js
@@ -1,3 +1,0 @@
-/** @entrypoint */
-import { ExportForm } from "./ExportForm";
-export default ExportForm;